refactor(toast): separate displayToast and dismissToast functions into their own modules
