+++ src/modules/toast/ui/dismissToast.ts
@@ -1,0 +1,8 @@
+import toast from 'solid-toast'
+
+/**
+ * Dismiss a toast by its solid-toast ID
+ */
+export function dismissToast(solidToastId: string): void {
+  toast.dismiss(solidToastId)
+}
